+++ controllers/location_controller.py
@@ -14,13 +14,17 @@ from schedule_app import app
 validator = LocationValidator()
 
 # here will be your code
-@app.route("/api/v1/location/<object_id>", methods=['PUT'])
+
+
+@app.route("/api/v1/location/<int:object_id>", methods=['PUT'])
 def update(object_id: int) -> Union[Tuple[str, int], Response]:
     """
     Обновляем Location
     :param object_id: int
     :return: Response
     """
+    if request.get_json().get('object_id') != object_id:
+        return "", 400
     try:
         validator.validate(request.get_json(), "PUT")
     except ValueError:
@@ -42,13 +46,15 @@ def get_locations() -> Response:
     return jsonify([i.__dict__() for i in Location.get_all(app.config.get("schedule_db_source"))])
 
 
-@app.route("/api/v1/location/<object_id>", methods=["GET"])
+@app.route("/api/v1/location/<int:object_id>", methods=["GET"])
 def get_location_by_id(object_id: int) -> Union[Response, Tuple[str, int]]:
     """
     Выдаём Location по заданному id
     :param object_id: int
     :return: Response
     """
+    if request.get_json().get('object_id') != object_id:
+        return "", 400
     try:
         return jsonify(Location.get_by_id(object_id, app.config.get("schedule_db_source")).__dict__())
     except ValueError:
@@ -69,13 +75,15 @@ def create_location() -> Response():
         return "", 400
 
 
-@app.route("/api/v1/location/<object_id>", methods=["DELETE"])
+@app.route("/api/v1/location/<int:object_id>", methods=["DELETE"])
 def delete_location(object_id: int) -> Union[Tuple[str, int], Tuple[Any, int], Response]:
     """
     Удаляем Location по заданному id
     :param object_id: int
     :return: Response
     """
+    if request.get_json().get('object_id') != object_id:
+        return "", 400
     try:
         location = Location.get_by_id(object_id, app.config.get("schedule_db_source"))
         location = location.delete().__dict__()
